child_process

+++ 10-file-system/03-my-mv.js
@@ -4,5 +4,10 @@ const fs = require('fs'),
       src = process.argv[2] || __filename,
       dst = process.argv[3];
 
-fs.renameSync(src,dst);
+if(fs.existsSync(src)){
+  fs.renameSync(src,dst);
+}else{
+  console.error('不存在');
+  process.exit(1);
+}
